+++ src/aigora/curriculum_graph/application/graph_loader.py
@@ -7,8 +7,8 @@ from aigora.curriculum_graph.domain.curriculum_graph import CurriculumGraph
 from aigora.curriculum_graph.application.assembling.graph_assembler import GraphAssembler
 from aigora.curriculum_graph.application.parsing.graph_parser import GraphParser
 from aigora.curriculum_graph.application.mapping.graph_mapper import GraphMapper
+from aigora.curriculum_graph.application.validation.graph_validator import GraphValidator
 
-from .graph_validator import GraphValidator
 from .loader_errors import GraphLoaderError
