+++ src/aigora/curriculum_graph/application/validation/graph_validator.py
