+++ src/aigora/curriculum_graph/application/validation/schema_errors.py
