+++ src/aigora/curriculum_graph/application/validation/validation_errors.py
